# backend/docs/swagger.yaml
definitions:
  gin_http.ResponseJSON:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/comment:
    post:
      parameters:
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: integer
      - description: 文章id
        in: formData
        name: article_id
        required: true
        type: integer
      - description: 回复评论id
        in: formData
        name: reply_id
        type: integer
      - description: 内容
        in: formData
        name: content
        required: true
        type: string
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 添加评论
  /api/v1/comment/{id}:
    delete:
      parameters:
      - description: 评论id
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 删除评论
  /api/v1/comment/{id}/like:
    post:
      parameters:
      - description: 评论id
        in: path
        name: id
        required: true
        type: integer
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: integer
      - description: 操作类型
        in: formData
        name: type
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 点赞评论
  /api/v1/comments/{article_id}:
    get:
      parameters:
      - description: 文章id
        in: path
        name: article_id
        required: true
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 获取评论
...
  /api/v1/search/tag:
    get:
      parameters:
      - description: tag_name
        in: query
        name: tag_name
        required: true
        type: string
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 标签补全
...
  /api/v1/tag:
    get:
      parameters:
      - description: tag_name
        in: query
        name: tag_name
        required: true
        type: string
      - description: user_id
        in: query
        name: uid
        type: integer
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 获取该标签的所有文章
  /api/v1/tag/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
        "10008":
          description: ""
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 删除标签
  /api/v1/tag/{id}/clear:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 清理标签(硬删除)
  /api/v1/tag/{id}/recover:
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 恢复标签
  /api/v1/tags:
    get:
      parameters:
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 获取标签列表
...
swagger: "2.0"